+++ kubejs/server_scripts/loot/blocks.js
@@ -1,0 +1,29 @@
+const registerBlockLoot = (/** @type {Internal.BlockLootEventJS} */ event) => {
+    let oreTypes = ['bauxite']
+    let gemTypes = ['quartz', 'clear_quartz']
+    global.tfcRockTypes.forEach(rock => {
+        let oreValues = ['poor', 'normal', 'rich']
+        oreValues.forEach(value => { 
+            let id = `cfc:ore/${value}_bauxite/${rock}`
+            event.addBlock(id, loot => {
+                loot.clearPools()
+                loot.addPool(pool => {
+                    pool.addItem(`cfc:ore/${value}_bauxite`)
+                })
+            })
+        })
+
+        event.addBlock(`cfc:ore/quartz/${rock}`, loot => {
+            loot.clearPools()
+            loot.addPool(pool => {
+                pool.addItem(`cfc:ore/quartz`)
+            })
+        })
+        event.addBlock(`cfc:ore/clear_quartz/${rock}`, loot => {
+            loot.clearPools()
+            loot.addPool(pool => {
+                pool.addItem(`cfc:ore/clear_quartz`)
+            })
+        })
+    })
+}