+++ kubejs/server_scripts/main_server.js
@@ -1,0 +1,26 @@
+// priority: 0
+
+ServerEvents.recipes(event => {
+    registerGeneralRecipes(event)
+})
+
+ServerEvents.tags('item', event => {
+    registerGeneralItemTags(event)
+})
+
+ServerEvents.tags('fluid', event => {
+    registerGeneralFluidTags(event)
+})
+
+ServerEvents.blockLootTables(event => {
+    registerBlockLoot(event)
+})
+
+TFCEvents.data(event => {
+    registerMetals(event)
+    registerHeats(event)
+    registerSizes(event)
+    registerLampFuels(event)
+    registerPanningData(event)
+    registerSluiceData(event)
+})