+++ kubejs/startup_scripts/events/glass_operations.js
@@ -1,0 +1,7 @@
+TFCEvents.createGlassOperations(event => {
+    event.createPowder('cfc:powder/bauxite', 'bauxite')
+    event.createPowder('cfc:powder/quartz', 'quartz')
+    event.createPowder('cfc:powder/cryolite', 'cryolite')
+    event.createPowder('cfc:powder/cinnabar', 'cinnabar')
+    event.createPowder('cfc:powder/chromite', 'chromite')
+})