+++ kubejs/startup_scripts/main_startup.js
@@ -1,0 +1,12 @@
+// priority: 0
+StartupEvents.registry('block', event => {
+    registerBlocks(event)
+})
+
+StartupEvents.registry('item', event => {
+    registerItems(event)
+})
+
+StartupEvents.registry('fluid', event => {
+    registerFluids(event)
+})