+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - 24-06-2024
+## [0.0.1] - 27-07-2025
 ### Changes
 - Compiled Modlist
+- Config Changes
+
+## [0.0.2] - 27-07-2025
+### Changes
+- New Items, Blocks & Fluids
+- New Glassblowing Powders
+- Basic TFC Data Setup
+- Player Event Setup